+++ sndgen/main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-	Channel c = Channel(8);
+	Channel c = Channel((size_t)8);
 	
 	for (auto i = 0; i < 100; i++)
 		c.push_sample(i);
